+++ src/ApplicationRunner.php
@@ -30,12 +30,14 @@ abstract class ApplicationRunner implements RunnerInterface
     /**
      * @param string $rootPath The absolute path to the project root.
      * @param bool $debug Whether the debug mode is enabled.
+     * @param string $paramsConfigGroup The config parameters group name.
      * @param string $containerConfigGroup The container configuration group name.
      * @param string|null $environment The environment name.
      */
     public function __construct(
         protected string $rootPath,
         protected bool $debug,
+        protected string $paramsConfigGroup,
         protected string $containerConfigGroup,
         protected ?string $environment
     ) {
@@ -163,7 +165,11 @@ abstract class ApplicationRunner implements RunnerInterface
      */
     protected function createDefaultConfig(): Config
     {
-        return ConfigFactory::create(new ConfigPaths($this->rootPath, 'config'), $this->environment);
+        return ConfigFactory::create(
+            new ConfigPaths($this->rootPath, 'config'),
+            $this->environment,
+            $this->paramsConfigGroup,
+        );
     }
 
     /**