+++ src/ConfigFactory.php
@@ -18,7 +18,7 @@ final class ConfigFactory
     /**
      * @throws ErrorException If the environment does not exist.
      */
-    public static function create(ConfigPaths $paths, ?string $environment): Config
+    public static function create(ConfigPaths $paths, ?string $environment, string $paramsGroup = 'params'): Config
     {
         $eventGroups = [
             'events',
@@ -33,6 +33,7 @@ final class ConfigFactory
                 ReverseMerge::groups(...$eventGroups),
                 RecursiveMerge::groups('params', ...$eventGroups),
             ],
+            $paramsGroup,
         );
     }
 }