+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - Сhg #37: Remove parameters in `ApplicationRunner` methods `runBootstrap()` and `checkEvents()`, instead are used 
   internal container and config instances (@vjik) 
-- Chg #36: Add ability to configure container configuration group (@vjik)
+- New #36: Add ability to configure container configuration group (@vjik)
+- New #38: Add ability to configure config params group (@vjik)
 
 ## 1.2.1 November 07, 2022